--- scripts/naf/données/liste_NAF_level2.json
@@ -1,204 +1,0 @@
-[
-	{
-		"code_NAF": "A01",
-		"libellé": "Culture et production animale, chasse et services annexes"
-	},
-	{ "code_NAF": "A02", "libellé": "Sylviculture et exploitation forestière" },
-	{ "code_NAF": "A03", "libellé": "Pêche et aquaculture" },
-	{ "code_NAF": "B05", "libellé": "Extraction de houille et de lignite" },
-	{ "code_NAF": "B06", "libellé": "Extraction d'hydrocarbures" },
-	{ "code_NAF": "B07", "libellé": "Extraction de minerais métalliques" },
-	{ "code_NAF": "B08", "libellé": "Autres industries extractives" },
-	{
-		"code_NAF": "B09",
-		"libellé": "Services de soutien aux industries extractives"
-	},
-	{ "code_NAF": "C10", "libellé": "Industries alimentaires" },
-	{ "code_NAF": "C11", "libellé": "Fabrication de boissons" },
-	{ "code_NAF": "C12", "libellé": "Fabrication de produits à base de tabac" },
-	{ "code_NAF": "C13", "libellé": "Fabrication de textiles" },
-	{ "code_NAF": "C14", "libellé": "Industrie de l'habillement" },
-	{ "code_NAF": "C15", "libellé": "Industrie du cuir et de la chaussure" },
-	{
-		"code_NAF": "C16",
-		"libellé": "Travail du bois et fabrication d'articles en bois et en liège, à l'exception des meubles ; fabrication d'articles en vannerie et sparterie"
-	},
-	{ "code_NAF": "C17", "libellé": "Industrie du papier et du carton" },
-	{
-		"code_NAF": "C18",
-		"libellé": "Imprimerie et reproduction d'enregistrements"
-	},
-	{ "code_NAF": "C19", "libellé": "Cokéfaction et raffinage" },
-	{ "code_NAF": "C20", "libellé": "Industrie chimique" },
-	{ "code_NAF": "C21", "libellé": "Industrie pharmaceutique" },
-	{
-		"code_NAF": "C22",
-		"libellé": "Fabrication de produits en caoutchouc et en plastique"
-	},
-	{
-		"code_NAF": "C23",
-		"libellé": "Fabrication d'autres produits minéraux non métalliques"
-	},
-	{ "code_NAF": "C24", "libellé": "Métallurgie" },
-	{
-		"code_NAF": "C25",
-		"libellé": "Fabrication de produits métalliques, à l'exception des machines et des équipements"
-	},
-	{
-		"code_NAF": "C26",
-		"libellé": "Fabrication de produits informatiques, électroniques et optiques"
-	},
-	{ "code_NAF": "C27", "libellé": "Fabrication d'équipements électriques" },
-	{
-		"code_NAF": "C28",
-		"libellé": "Fabrication de machines et équipements n.c.a."
-	},
-	{ "code_NAF": "C29", "libellé": "Industrie automobile" },
-	{
-		"code_NAF": "C30",
-		"libellé": "Fabrication d'autres matériels de transport"
-	},
-	{ "code_NAF": "C31", "libellé": "Fabrication de meubles" },
-	{ "code_NAF": "C32", "libellé": "Autres industries manufacturières" },
-	{
-		"code_NAF": "C33",
-		"libellé": "Réparation et installation de machines et d'équipements"
-	},
-	{
-		"code_NAF": "D35",
-		"libellé": "Production et distribution d'électricité, de gaz, de vapeur et d'air conditionné"
-	},
-	{ "code_NAF": "E36", "libellé": "Captage, traitement et distribution d'eau" },
-	{ "code_NAF": "E37", "libellé": "Collecte et traitement des eaux usées" },
-	{
-		"code_NAF": "E38",
-		"libellé": "Collecte, traitement et élimination des déchets ; récupération"
-	},
-	{
-		"code_NAF": "E39",
-		"libellé": "Dépollution et autres services de gestion des déchets"
-	},
-	{ "code_NAF": "F41", "libellé": "Construction de bâtiments" },
-	{ "code_NAF": "F42", "libellé": "Génie civil" },
-	{ "code_NAF": "F43", "libellé": "Travaux de construction spécialisés" },
-	{
-		"code_NAF": "G45",
-		"libellé": "Commerce et réparation d'automobiles et de motocycles"
-	},
-	{
-		"code_NAF": "G46",
-		"libellé": "Commerce de gros, à l'exception des automobiles et des motocycles"
-	},
-	{
-		"code_NAF": "G47",
-		"libellé": "Commerce de détail, à l'exception des automobiles et des motocycles"
-	},
-	{
-		"code_NAF": "H49",
-		"libellé": "Transports terrestres et transport par conduites"
-	},
-	{ "code_NAF": "H50", "libellé": "Transports par eau" },
-	{ "code_NAF": "H51", "libellé": "Transports aériens" },
-	{
-		"code_NAF": "H52",
-		"libellé": "Entreposage et services auxiliaires des transports"
-	},
-	{ "code_NAF": "H53", "libellé": "Activités de poste et de courrier" },
-	{ "code_NAF": "I55", "libellé": "Hébergement" },
-	{ "code_NAF": "I56", "libellé": "Restauration" },
-	{ "code_NAF": "J58", "libellé": "Édition" },
-	{
-		"code_NAF": "J59",
-		"libellé": "Production de films cinématographiques, de vidéo et de programmes de télévision ; enregistrement sonore et édition musicale"
-	},
-	{ "code_NAF": "J60", "libellé": "Programmation et diffusion" },
-	{ "code_NAF": "J61", "libellé": "Télécommunications" },
-	{
-		"code_NAF": "J62",
-		"libellé": "Programmation, conseil et autres activités informatiques"
-	},
-	{ "code_NAF": "J63", "libellé": "Services d'information" },
-	{
-		"code_NAF": "K64",
-		"libellé": "Activités des services financiers, hors assurance et caisses de retraite"
-	},
-	{ "code_NAF": "K65", "libellé": "Assurance" },
-	{
-		"code_NAF": "K66",
-		"libellé": "Activités auxiliaires de services financiers et d'assurance"
-	},
-	{ "code_NAF": "L68", "libellé": "Activités immobilières" },
-	{ "code_NAF": "M69", "libellé": "Activités juridiques et comptables" },
-	{
-		"code_NAF": "M70",
-		"libellé": "Activités des sièges sociaux ; conseil de gestion"
-	},
-	{
-		"code_NAF": "M71",
-		"libellé": "Activités d'architecture et d'ingénierie ; activités de contrôle et analyses techniques"
-	},
-	{ "code_NAF": "M72", "libellé": "Recherche-développement scientifique" },
-	{ "code_NAF": "M73", "libellé": "Publicité et études de marché" },
-	{
-		"code_NAF": "M74",
-		"libellé": "Autres activités spécialisées, scientifiques et techniques"
-	},
-	{ "code_NAF": "M75", "libellé": "Activités vétérinaires" },
-	{ "code_NAF": "N77", "libellé": "Activités de location et location-bail" },
-	{ "code_NAF": "N78", "libellé": "Activités liées à l'emploi" },
-	{
-		"code_NAF": "N79",
-		"libellé": "Activités des agences de voyage, voyagistes, services de réservation et activités connexes"
-	},
-	{ "code_NAF": "N80", "libellé": "Enquêtes et sécurité" },
-	{
-		"code_NAF": "N81",
-		"libellé": "Services relatifs aux bâtiments et aménagement paysager"
-	},
-	{
-		"code_NAF": "N82",
-		"libellé": "Activités administratives et autres activités de soutien aux entreprises"
-	},
-	{
-		"code_NAF": "O84",
-		"libellé": "Administration publique et défense ; sécurité sociale obligatoire"
-	},
-	{ "code_NAF": "P85", "libellé": "Enseignement" },
-	{ "code_NAF": "Q86", "libellé": "Activités pour la santé humaine" },
-	{ "code_NAF": "Q87", "libellé": "Hébergement médico-social et social" },
-	{ "code_NAF": "Q88", "libellé": "Action sociale sans hébergement" },
-	{
-		"code_NAF": "R90",
-		"libellé": "Activités créatives, artistiques et de spectacle"
-	},
-	{
-		"code_NAF": "R91",
-		"libellé": "Bibliothèques, archives, musées et autres activités culturelles"
-	},
-	{
-		"code_NAF": "R92",
-		"libellé": "Organisation de jeux de hasard et d'argent"
-	},
-	{
-		"code_NAF": "R93",
-		"libellé": "Activités sportives, récréatives et de loisirs"
-	},
-	{ "code_NAF": "S94", "libellé": "Activités des organisations associatives" },
-	{
-		"code_NAF": "S95",
-		"libellé": "Réparation d'ordinateurs et de biens personnels et domestiques"
-	},
-	{ "code_NAF": "S96", "libellé": "Autres services personnels" },
-	{
-		"code_NAF": "T97",
-		"libellé": "Activités des ménages en tant qu'employeurs de personnel domestique"
-	},
-	{
-		"code_NAF": "T98",
-		"libellé": "Activités indifférenciées des ménages en tant que producteurs de biens et services pour usage propre"
-	},
-	{
-		"code_NAF": "U99",
-		"libellé": "Activités des organisations et organismes extraterritoriaux"
-	}
-]
