+++ scripts/services-societaux/input/SDES_groups_to_split.json
@@ -1,0 +1,131 @@
+{
+	"B05_B09": {
+		"B05": {
+			"libellé": "Extraction de houille et de lignite"
+		},
+		"B06": { "libellé": "Extraction d'hydrocarbures" },
+		"B07": {
+			"libellé": "Extraction de minerais métalliques"
+		},
+		"B08": {
+			"libellé": "Autres industries extractives"
+		},
+		"B09": {
+			"libellé": "Services de soutien aux industries extractives"
+		}
+	},
+	"C10_C12": {
+		"C10": {
+			"libellé": "Industries alimentaires"
+		},
+		"C11": {
+			"libellé": "Fabrication de boissons"
+		},
+		"C12": {
+			"libellé": "Fabrication de produits à base de tabac"
+		}
+	},
+	"C13_C15": {
+		"C13": { "libellé": "Fabrication de textiles" },
+		"C14": {
+			"libellé": "Industrie de l'habillement"
+		},
+		"C15": {
+			"libellé": "Industrie du cuir et de la chaussure"
+		}
+	},
+	"C31_C32": {
+		"C31": { "libellé": "Fabrication de meubles" },
+		"C32": {
+			"libellé": "Autres industries manufacturières"
+		}
+	},
+	"E37_E39": {
+		"E37": {
+			"libellé": "Collecte et traitement des eaux usées"
+		},
+		"E38": {
+			"libellé": "Collecte, traitement et élimination des déchets ; récupération"
+		},
+		"E39": {
+			"libellé": "Dépollution et autres services de gestion des déchets"
+		}
+	},
+	"F41_F43": {
+		"F41": {
+			"libellé": "Construction de bâtiments"
+		},
+		"F42": { "libellé": "Génie civil" },
+		"F43": {
+			"libellé": "Travaux de construction spécialisés"
+		}
+	},
+	"I55_I56": {
+		"I55": { "libellé": "Hébergement" },
+		"I56": { "libellé": "Restauration" }
+	},
+	"J59_J60": {
+		"J59": {
+			"libellé": "Production de films cinématographiques, de vidéo et de programmes de télévision ; enregistrement sonore et édition musicale"
+		},
+		"J60": {
+			"libellé": "Programmation et diffusion"
+		}
+	},
+	"J62_J63": {
+		"J62": {
+			"libellé": "Programmation, conseil et autres activités informatiques"
+		},
+		"J63": { "libellé": "Services d'information" }
+	},
+	"M69_M70": {
+		"M69": {
+			"libellé": "Activités juridiques et comptables"
+		},
+		"M70": {
+			"libellé": "Activités des sièges sociaux ; conseil de gestion"
+		}
+	},
+	"M74_M75": {
+		"M74": {
+			"libellé": "Autres activités spécialisées, scientifiques et techniques"
+		},
+		"M75": { "libellé": "Activités vétérinaires" }
+	},
+	"N80_N82": {
+		"N80": { "libellé": "Enquêtes et sécurité" },
+		"N81": {
+			"libellé": "Services relatifs aux bâtiments et aménagement paysager"
+		},
+		"N82": {
+			"libellé": "Activités administratives et autres activités de soutien aux entreprises"
+		}
+	},
+	"Q87_Q88": {
+		"Q87": {
+			"libellé": "Hébergement médico-social et social"
+		},
+		"Q88": {
+			"libellé": "Action sociale sans hébergement"
+		}
+	},
+	"R90_R92": {
+		"R90": {
+			"libellé": "Activités créatives, artistiques et de spectacle"
+		},
+		"R91": {
+			"libellé": "Bibliothèques, archives, musées et autres activités culturelles"
+		},
+		"R92": {
+			"libellé": "Organisation de jeux de hasard et d'argent"
+		}
+	},
+	"T97_T98": {
+		"T97": {
+			"libellé": "Activités des ménages en tant qu'employeurs de personnel domestique"
+		},
+		"T98": {
+			"libellé": "Activités indifférenciées des ménages en tant que producteurs de biens et services pour usage propre"
+		}
+	}
+}
